# asklepios-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asklepios ERP API Documentation",
        "description": "Documentation interactive de l'API Asklepios pour la gestion hospitalière. Permet aux développeurs front-end de tester les requêtes en temps réel.",
        "contact": {
            "name": "Support Technique Asklepios",
            "email": "admin@asklepios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API de Développement"
        }
    ],
    "paths": {
        "/api/admin/article-categories": {
            "get": {
                "tags": [
                    "Catégories d'Articles (Admin)"
                ],
                "summary": "Lister les catégories d'articles (Paginées)",
                "description": "Lister les catégories (Paginées, pour le tableau d'affichage)",
                "operationId": "getAdminArticleCategories",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de résultats par page (défaut: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page à récupérer (défaut: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catégories d'Articles (Admin)"
                ],
                "summary": "Créer une catégorie",
                "description": "Créer une nouvelle catégorie",
                "operationId": "storeAdminArticleCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Antibiotiques"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Médicaments antibactériens"
                                    },
                                    "article_category_id": {
                                        "description": "ID de la catégorie parente",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/article-categories/all": {
            "get": {
                "tags": [
                    "Catégories d'Articles (Admin)"
                ],
                "summary": "Lister toutes les catégories (Sans pagination)",
                "description": "Lister TOUTES les catégories (Sans pagination, pour les select du frontend)",
                "operationId": "getAllAdminArticleCategories",
                "responses": {
                    "200": {
                        "description": "Liste complète récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/article-categories/{id}": {
            "put": {
                "tags": [
                    "Catégories d'Articles (Admin)"
                ],
                "summary": "Modifier une catégorie",
                "description": "Modifier une catégorie",
                "operationId": "updateAdminArticleCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "article_category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie modifiée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catégories d'Articles (Admin)"
                ],
                "summary": "Supprimer une catégorie",
                "description": "Supprimer une catégorie",
                "operationId": "deleteAdminArticleCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Saisissez le jeton (Token) obtenu lors de la connexion pour tester les routes protégées.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}